+++ stroom-security/stroom-security-identity/src/main/java/stroom/security/identity/exceptions/BadRequestException.java
@@ -18,15 +18,17 @@
 
 package stroom.security.identity.exceptions;
 
+import event.logging.AuthenticateOutcomeReason;
+
 import javax.annotation.Nullable;
 
 public class BadRequestException extends RuntimeException {
     private final String subject;
 
     //See AuthenticateOutcomeReason for known values
-    private final String reason;
+    private final AuthenticateOutcomeReason reason;
 
-    public BadRequestException(@Nullable final String subject, final String reason, final String message) {
+    public BadRequestException(@Nullable final String subject, final AuthenticateOutcomeReason reason, final String message) {
         super(message);
         this.subject = subject;
         this.reason = reason;
@@ -36,7 +38,7 @@ public class BadRequestException extends RuntimeException {
         return subject;
     }
 
-    public String getReason() {
+    public AuthenticateOutcomeReason getReason() {
         return reason;
     }
 }